pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "slop-bot"
version = "0.1.0"
description = "A Discord bot project"
authors = [{name = "Your Name", email = "you@example.com"}]
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.11"
dependencies = [
    "discord.py==2.3.2",
    "httpx==0.25.2",
    "python-dotenv==1.0.0",
    "pydantic==2.8.0",
    "Pillow==10.0.1",
    "aiofiles==24.1.0",
    "requests==2.32.0",
    "uvicorn[standard]==0.30.0",
    "fastapi==0.115.0"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12"
]
[project.optional-dependencies]
dev = [
    "pyright==1.1.378",
    "pytest==8.0.0",
    "pytest-asyncio==0.23.0"
]
[project.scripts]
slop-bot = "src.bot:main"
